crunchdjango/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.238
botocore==1.12.238
celery==4.3.0
Django==2.2.5
django-cors==0.1
django-cors-headers==3.1.0
djangorestframework==3.10.3
docutils==0.15.2
importlib-metadata==0.22
jmespath==0.9.4
kombu==4.6.4
more-itertools==7.2.0
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
